+++ src/common/SortedOutputBox.jsx
@@ -1,36 +1,7 @@
 import styled from "@emotion/styled";
+import useSort from "utils/hooks/useSort.js";
 
 const SortedOutputBox = ({ numbers, direction }) => {
-  const arrangeNumbers = () => {
-    const arrangedNumberArr =
-      direction === "ascending"
-        ? arrangeAscendingOrder(numbers)
-        : arrangeDescendingOrder(numbers);
-    return addCommasToNumbers(arrangedNumberArr);
-  };
-
-  const arrangeAscendingOrder = (arr) => {
-    for (let i = 0; i < arr.length; i++) {
-      for (let j = i + 1; j < arr.length; j++) {
-        if (arr[i] > arr[j]) {
-          [arr[i], arr[j]] = [arr[j], arr[i]];
-        }
-      }
-    }
-    return arr;
-  };
-
-  const arrangeDescendingOrder = (arr) => {
-    for (let i = 0; i < arr.length; i++) {
-      for (let j = i + 1; j < arr.length; j++) {
-        if (arr[i] < arr[j]) {
-          [arr[i], arr[j]] = [arr[j], arr[i]];
-        }
-      }
-    }
-    return arr;
-  };
-
   const addCommasToNumbers = (arr) => {
     return arr.map((num, idx) => {
       return idx < arr.length - 1 ? `${num}${", "}` : `${num}`;
@@ -39,7 +10,9 @@ const SortedOutputBox = ({ numbers, direction }) => {
 
   return (
     <SortedContainer>
-      <SortedLayout>{arrangeNumbers()}</SortedLayout>
+      <SortedLayout>
+        {addCommasToNumbers(useSort({ numbers, direction }))}
+      </SortedLayout>
     </SortedContainer>
   );
 };
